encapp: Refactorization and update to configuraton and SetParameter
For easier test case handling an new way of running and defining tests
is in place.
Test are created as json files and pushed to the device.
Testname is flagged with "-e test PATH" where path is
path to json file with test description(s).

A few tests are defined in /tests/ and corresponding videos in /media/