encapp: Added 'realtime' and 'pursuit' mode
With realtime mode the decoded (or raw) frames will be delivered to the encoder
at the correct wall time according to the settings and the system works as if had
a live source for the video.

In pursuit mode encapp will create a new encoder every second until it fails to do so.
Then it will run all running encoders to finish and report their performance.