encapp: Decode only mode and pursuit mode changes
* A new mode with no encode to test decoder support.
	"encode": "false"

* Pursuit mode has changes so it can be created with a spcific number of
decoders or until it failes:
	-1 - create decoders/encoders until failure
	0  - no pursuit mode
	X  - create X number of encoders/decoders
exampel: "pursuit" : "24"
means create 24 instances of the test

* Pursuit mode has an end criteria.
Pursuit mde ends when a test fails to initiate and no also when the first test is finished
(Encapp stops creating more instances).

* Create specified decoder
	"decoder": "OMX.qcom.video.decoder.avc"
If this is not specified the mime type of the decoded file will be used to create decoder.

* a minor fix for input file override which caused a problem in the encapp.py script.