encappng: replace "duration" with "playout-frames"
Offline discussion.

There was a comment about "duration" being the wrong word to express that
we want to play 990 frames (30 seconds of 30 fps video) from a file that
is only 20 frames. This field is used to play out a given number of frames
from the input, different from the actual duration. Therefore we are using
"playout-frames" instead of "duration".