encapp: vmaf must be called with videos in the right order
VMAF with ffmpeg expects to be called with the video streams in the
right order:

```
$ ffmpeg <distorted_video> <reference_video> ...
```

If the order is reversed, the scores are completely broken. For example,
a quick experiment encoding a 1080p using x265 at 2 Mbps (default x265
parameters), show a 60.489530 score when the videos are passed in the
right order, and a 99.844539 score when the videos are passed in the
wrong order.

Reference:

https://jina-liu.medium.com/a-practical-guide-for-vmaf-481b4d420d9c