encapp: add a generic, CLI-based, variable replace mechanism
Idea is to allow changing all values in the protobuf by naming them
deterministically. For example, in order to change the value of
test['foo'][bar'] to "baz", the user will add "-e foo.bar baz".
Thus, we have "-e input.filepath baz" to change the input filepath
to "baz". Same way, we have "-e configure.bitrate baz" for the
encoder bitrate.

We add a couple of shortcuts for CLI parameters that we use a lot,
including:

* "-i" (videofile) for input.filepath
* "-c/--codec" for configure.codec
* "-r/--bitrate" for configure.bitrate