encapp: fix start option
A left over from old code. The -w was used to wait for test finished.
At one point some DUT just ignored this and a new mechanism for
checking finsiehd tests were introduced, It turns out that some
devices can hang here and since it is not needed - it is removed.

Signed-off-by: JohanBlome <jblome@fb.com>