encapp: fix for forcing raw input
Added a new cli option to force raw transcode
"--raw".
The old way is a mess of special handling to accomodate specific test
situations. It needs to be simplified. In the meantime this cli will
force raw transcode to unblock e.g. ios.

A bug with the stride calculation is also fixed.

Signed-off-by: JohanBlome <jblome@fb.com>