encapp: add os walk and regexp search to video file path
input.filepath can now be either:
1) specific file
2) a folder
3) a root folder with a regexp filename

A file walk will be performed at the root folder (last '/')
and files will be matched based on endings and a regexp filter.
E.g
    input {
        filepath: "~/Downloads/.*[0-1]{1}.mov"
    }

Matching anythign in the downloads folder with a single digit of either 0 or 1
and '.mov' ending.

It will traverse subfolders as well as the root.

Signed-off-by: JohanBlome <jblome@fb.com>