encapp: refactorization of codecslist
1)
Codec list is not a proper json.
The function 'list' can now be used for querying and result is cached.

>python3 encapp.py list -c ".*hevc"
c2.exynos.hevc.encoder
c2.android.hevc.encoder

The codec search takes a regexp (if using a plain wildcard it will handle that as well).
The output can be named and exported to a specific location and not just in the active directory.

For all arguments run:
>python3 encapp.py list -h

2)
System test is updated with some basic tests.
The normal way to run it will include a installation of latest build.
This can be overridden to test the current installation on the device.

Signed-off-by: JohanBlome <jblome@fb.com>