encapp: remove corpus test
The corpus test is not needed anymore. Everything it could do can be done
by encapp.py i.e. run a complete test over a folder.

Either run a very small or empty protobuf file or provide a "." as
protobuf filename and a temporary will be created.
The input file can now be a folde as well as a file.

Note and TODO:
   * "-i" does not accept regex, currently only actual file/folders are possibleI

In a protobuf file the filepath can be a regex so that is more flexible.

e.g:
>pyton3 encapp.py run .  -c c2.exynos.hevc.encoder -r 250kbps,500kbps  --quality -i corpus

Signed-off-by: JohanBlome <jblome@fb.com>