encapp: add options for VMAF partial frame calculations
By default the VMAF calculation will be run on the same resolution as the source.
With
"--vmaf-crop WxH" a cropped part of the video can be used for the calculation.
It will be a center crop (todo: a offset is added in the args but not implemented)

As an alternative both the original and the distorted fiel can be scaled to a common resolution:
"--vmaf-scale WxH"

Default scaler is lanczos but it can be set through:
"--vmaf-scaler XXX"

Signed-off-by: Johan Blome <jblome@meta.com>